refactor: Move the difficulty tests into their own file
